requirements.txt
asgiref==3.8.1
cosine-similarity==0.1.2
Django==4.2.17
djangorestframework==3.15.2
joblib==1.4.2
numpy==1.26.4
pandas==2.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.6.1
scikit-surprise==1.1.4
scipy==1.13.1
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1

celery~=5.4.0